Make button appearance variant reusable
