+++ packages/nimble-components/src/button/index.ts
@@ -5,7 +5,7 @@ import {
     buttonTemplate as template,
     DesignSystem
 } from '@microsoft/fast-foundation';
-import type { ButtonPattern } from '../patterns/button/types';
+import type { ButtonPattern, ButtonWithAppearanceVariantPattern } from '../patterns/button/types';
 import { styles } from './styles';
 import { ButtonAppearance, ButtonAppearanceVariant } from './types';
 
@@ -18,7 +18,7 @@ declare global {
 /**
  * A nimble-styled HTML button
  */
-export class Button extends FoundationButton implements ButtonPattern {
+export class Button extends FoundationButton implements ButtonPattern, ButtonWithAppearanceVariantPattern {
     /**
      * @public
      * @remarks
